feat: add ApplyOptions + move to Jest (#19)
* feat: add ApplyOptions

lets face it, we'll be on our deathbeds by the time Jacz would do this
himself. As for testing the decorator, for the life of me I've tried
but I can't wrap my head around all the mocking required just to get it
to work. My idea was to have a class which would have a method that
returns one of its properties (such as cooldown for a command) but then
that requires a KlasaGuild, KlasaMessage, TextChannel, Channel,
KlasaClient and god knows what more.

Signed-off-by: Jeroen Claassens <support@favware.tech>

* Update src/index.ts

Co-Authored-By: Antonio Román <kyradiscord@gmail.com>

* Update src/index.ts

Co-Authored-By: Antonio Román <kyradiscord@gmail.com>

* feat: massive overhaul

- switched tape for jest. This is primarily because tape is just
annoying to work with but also because Jest is a way more mature
framework in terms of community and "market" <_< (it has nigh everything

- Implemented tests for ApplyOptions

- Updated azure pipelines config (ubuntu 18 + jest coverage)

- Moved tests outside of src -> caveat is that there is now a
tsconfig.build.json for the building process. __tests__ needs to be
included because otherwise VSC is tripping over the decorators but
__tests__ should not be in the dist, that's ugly af. This is the best
solution.

- Removed npmignore, better to specificy which few files to include
rather than which many files to exclude (see files field in
package.json)

- Updated eslint config to know we're using Jest

* ci: fix azure pipelines (hopefully)

* test: remove unnecessary await

* ci: maybe this fixes azure pipelines

* ci: revert back to ubutnu 16

* ci: switch to github workflows