feat(*): rewrite to DJS v12 + Klasa fork
As of this version this package is back to working with DiscordJS, with the major caveat that it
will only work with skyra-project/klasa fork. This is a transitional release until Skyra has moved
to Sapphire

BREAKING CHANGE: This package is no longer designed to be used for Klasa Core, which has been
discontinued. Furthermore, as Klasa as a whole has been discontinued, we do not recommend using this
package in your own projects. Instead you should be looking towards changing to the Sapphire
framework (https://github.com/sapphire-project) and using its package, @sapphire/decorators, which
has been greatly inspired by, and made by the same developers as @skyra/decorators.