fix(*): migrate to i18next
This completely removes support for the old Klasa based Languages and lots of Klasa related stuff.
Do not use this if you're not a Skyra developer. Migrate to Sapphire Framework instead and use
@sapphire/decorators.

BREAKING CHANGE: This completely removes support for the old Klasa based Languages and lots of Klasa
related stuff. Do not use this if you're not a Skyra developer. Migrate to Sapphire Framework
instead and use @sapphire/decorators.